Updates so it now mostly works.
Still can't sync properly with Google Drive. The per site stipulation isn't working properly. And the "plain" canned response type is showing both a plain window and the TinyMCE editor. When selecting the "HTML" option only the TinyMCE editor is showing.

The good news is that the copy function works properly and works for both plain text and HTML